% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{neat_chart}
\alias{neat_chart}
\title{neat_chart}
\usage{
neat_chart()
}
\value{
none
}
\description{
a function that cleans up plot appearance by closing the axis gap
}
\examples{
a_plot + neat_chart()
}
