% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_colors.R
\name{pastels1}
\alias{pastels1}
\title{pastels1}
\usage{
pastels1(color_name)
}
\arguments{
\item{color_name}{}
}
\value{
color name and hex code
}
\description{
5 soft, distinct pastel colors apt for graphing.
}
\examples{
ggplot(aes(x= xVariable, y = yVariable)) + geom_col(fill = pastels1("beige"))
}
